+++ Changelog.md
@@ -2,7 +2,7 @@
 
 ## v3.6.0
 
-- Promote C/C++ features to general availability [#1085](https://github.com/fossas/fossa-cli/pull/1085).
+- Promote C/C++ features to general availability [#1087](https://github.com/fossas/fossa-cli/pull/1087).
   - `--experimental-enable-vsi` is now `--detect-vendored`
   - `--experimental-analyze-dynamic-deps` is now `--detect-dynamic`.